container-telegram/src/requirements.txt
environs
envs
redis
python-telegram-bot
